requirements.txt
Flask
opencv-python
pillow
ultralytics
Flask-WTF
mathplotlib